+++ Changelog.md
@@ -8,6 +8,7 @@ Version 0.1.0 (Unreleased)
 
 Features
 --------
+- Implement Thrift-based binary injection.
 - Initial library structure.
 - TraceID implementation.
 - Zipkin SpanContext implementation.